api_pytest/requirements.txt
allure_python_commons==2.13.2
jsonpath==0.82.2
pymysql==1.1.0
pytest==7.1.2
PyYAML==6.0
PyYAML==6.0.1
Requests==2.31.0
